refactor : .client 네이밍 컨벤션
