+++ src/app/join/funnels/step3/components/EmailForm.client.tsx
@@ -7,7 +7,7 @@ import { useJoinContext } from '@/app/join/components/JoinContext';
 import { SignUpState } from '@/app/join/type';
 import BottomFixedDiv from '@/components/common/BottomFixedDiv';
 import { Button } from '@/components/common/Button';
-import { useModalContext } from '@/components/common/Modal/ModalContext';
+import { useModalContext } from '@/components/common/Modal/ModalContext.client';
 import { Spacing } from '@/components/common/Spacing';
 import { memo } from 'react';
