+++ src/components/common/Modal/index.ts
@@ -1,3 +1,3 @@
 export { default as ModalWrapper } from './ModalWrapper';
-export { default as BottomSheet } from './BottomSheet';
-export { default as ConfirmModal } from './ConfirmModal';
+export { default as BottomSheet } from './BottomSheet.client';
+export { default as ConfirmModal } from './ConfirmModal.client';
